feature: better theming backend support - replaces static_views system with better theming with overrides
long version: this modifies the way ubersystem searches for templates in the themed/ dir

here's the search order for the file:   'themed/prereg.css'
- first looks in 'themed/{current_theme_name}/prereg.css
- if not found, looks in 'themed/common/prereg.css'

Now we can override any file in common with stuff in specific theme dirs.

In the future, I want to unify this with the static/ dir, but for now this will work.  To unify both of them means we have to override more than just the templates and know when to process templates vs when to just serve up non-templated files (like JPGs and PNGs)