allow override of core enums
There's gotta be a better way to do this, but for now create some overriden python modules based on what theme is selected.

These are in siteconfig and are for database fields like shirt size, staffing area, areas of interest.

What's going on behind the curtain is a little hacky, but I think it's OK.  We need to probably revisit how we populate and match these enums from the DB to be more config-file friendly.