WIP shirt type + color selection
This deviates a bit from how shirt stuff was setup before in ubersystem, in Magstock it's based 100% on attendee.amount_extra

Since there are no other badge types, supporter badges, or donations, the only thing that can be in amount_extra is the shirt cost.

Note: that we are adding a new DB field here, attendee.shirt_color which can be 'black' or 'white for tie die'

We'll probably also need to create some kind of report that shows the breakdown of each shirt type.  When doing that, base the shirt counts ONLY on attendee.shirt != 0