Fix badges dropdown on group admin page
The badges dropdown on the group admin page used badges_purchased, but
the logic of the code for the groups form treated it as the total number
of badges. That meant you would have cases where if you simply tried to
save a group's info, it would actually try to reduce the number of
badges in the group. This fixes that, and adds a badges_purchased count
next to the dropdown.