Update configspec.ini
This changes the order of the Volunteer Checklist. We tell staffers/volunteers  that they have to do the steps in order... so that the next step actives.   During the registration process, or badge confirmation process people are already able to add their shirt size so that completes that step..... Making that step 2 instead of 3 shows the staffer/volunteer that they have already completed 2 steps in the process.