Changes: c.event shifts available subject line.
I am not sure if we want this change to be here or in the new magfest/magfest email section. 

This is a change to the Staff Operations email that goes out to all volunteers/staffers. 
The original iteration of the email was simply for volunteers to sign up for this shifts hence the "EVENT_NAME shifts available" subject line. 

For MAGFest events that email and the process has turned into a 4 or 5 step checklist (depending on if the person is a volunteer or staffer).   The MAGFest Staff Operations department refers to this Email notification and the process it entails as the Volunteer Checklist in all communication to staffers (Email, slack, fb.)   

This edit would chance the subject line of the email to: "Please complete your {EVENT_NAME} Staff/Volunteer Checklist"