Hide shirt levels after shirt_deadline to fix at-door kickin display (#2247)
Fixes https://github.com/magfest/ubersystem/issues/2224 and fixes https://github.com/magfest/ubersystem/issues/2210 so that shirt_deadline is actually meaningful. Also adds shirt_count so it mirrors the way the supporter count works. Slightly tweaks the kick-in display so that it will be hidden for attendees who have not already kicked in if there's only one option (because that option is always 'no thanks').