Updated closing date. (#2248)
* Updated closing date.

Emailed Code about this - but this email was linked to uber_takedown variable, which is set to be the middle of the event. 
I think that's old code related to when Uber was taken offline during load-in.
I have updated the link here to be Event Epoch (Start)

* Use final_email_deadline instead of epoch

The FINAL_EMAIL_DEADLINE var returns either uber_takedown OR epoch, whichever comes first. That's precisely the behavior we want here.