Always return date-based badge price if it's between epoch and eschaton (#2246)
* Always return date-based badge price if it's between epoch and eschaton
Fixes https://github.com/magfest/ubersystem/issues/2235. Also moves around the badges_sold calculation so we only poll the DB if we need to.

* Move badges_sold calculation back under optimizing if statement

I completely goofed on moving one line when the entire if statement should have been moved. Also fixes a UTC/localized time bug that we never encountered because we were always setting dt (and now we are not).

* Fix pep8 and clean up code