Implement QR Codes for Check-In (#2265)
* Generate unencrypted barcodes on confirm page
This is one of the first steps to server-side barcode generation. We may not go in this direction, however.

* Generate unencrypted barcodes on confirm page
This is one of the first steps to server-side barcode generation. We may not go in this direction, however.

* Fully functioning QR codes with encryption
This is the first pass for a QR code check-in system. It's missing a few things, but all the basics are there.

* Fix pep8 comma complaint
Huh, I did not know this was a thing.

* Remove extra barcode_generator function
This was a merge error.

* Only try to generate barcode image if QR_CODE_PASSWORD exists
Otherwise we'll get broken QR code images if this config option is not set.

* Send email with QR code to attendees who can check in
Although it's on their confirmation page, most attendees won't know to look for it there, so we send them the same image directly via email.

* Use a faster QR code encryption algorithm
This removes simple-crypt in favor of using its dependency, pycrypto, to encrypt the UUID without such a long delay. Thanks, Eli!