Add unique 'search key' for attendees and groups (#2344)
* Add unique attendee 'search key'
Fixes https://github.com/magfest/ubersystem/issues/2343 by adding a new UUID used only for search. Changes the check-in barcodes to use that UUID, preventing all security issues with using the primary key id. Also removes the encryption that is no longer needed.

* Re-add auto-creation of EVENT_QR_ID

This was removed by accident.

* Add USE_CHECKIN_BARCODE
Events without barcode scanners may not want checkin barcodes displayed. Since we got rid of QR_CODE_PASSWORD, we need another variable to control this.

* Don't assume anything starting with EVENT_QR_ID is a uuid
This hides the search by UUID under an if statement, so that if the string after EVENT_QR_ID isn't actually a UUID, we search as if it's normal text.

* Rename search_key to public_id