Add date to some email subjects (#1593)
* add dates to certain email subject lines
- ideally we would replace ALL {EVENT_NAME}, but doing it mid-year (for Magfest) would cause big issues with emails being re-sent.

* fix pep8

* change {EVENT_NAME_AND_DATE} to {EVENT}
remove c.EVENT_NAME_AND_DATE, put the string build inside the automated emails code because c.EVENT is not super-clear what it's supposed to be doing

* Change format of date to just month+year

It may be confusing to list the day (recipients may take that to mean it's a one-day event), and briefer is better. Also adds parentheses.