Expand definition of 'dealer' for the purpose of badge prices (#2393)
* Expand definition of 'dealer' for the purpose of badge prices
Fixes https://github.com/magfest/magclassic/issues/52. When we changed the way group prices worked, we started inspecting attendees' status for their price instead of the group's status. This definition wasn't broad enough for dealers, who could be in a 'dealer' group but not have a Shopkeep ribbon. This fixes that, and changes our dealer ribbon auto-assignment so it doesn't override admin choices by accident now that we've broadened this definition. It also auto-assigns Shopkeep ribbons to replacement attendees when you 'unassign' an attendee, which is the bug that revealed this problem.

* Clean up code and update test

* Actually test for dealership