Remove nonfunctional "Scan 2D Barcode" button
This was an oversight in https://github.com/magfest/ubersystem/pull/2344 -- we removed the `qr_code_reader` function that this button depends on, but not the button itself. Trying to use this button would result in an error.

Since #2344 lets you scan the check-in barcode directly into the searchbox, we don't need the button anymore.