remove some secrets
remove stripe keys and some magfest-specific settings in configspec.ini

These aren't particularly sensitive but, probably don't belong here.  In particular, magfest's stripe key (which is not a big deal because it's a test key so it can't any financial harm), but could be used for a denial of service attack or something.   Also, we're about to rotate it so it is going to be invalid anyway