Fix emails never including a link to the consent form
Fixes https://github.com/magfest/ubersystem/issues/2493. During testing, MFF found that confirmation emails were not actually including the consent form link for attendees under 18. It turned out to be because we had the syntax wrong for checking whether the attendee's age group needed a consent form. This fixes that.