Fix 2392 (#2514)
* Stop normalizing the left side of email addresses.
Fixes https://github.com/magfest/ubersystem/issues/2392. We don't need to normalize the left side of emails, so now we just don't. We DO normalize the domain, so we now also normalize it when someone is trying to log in -- that way we can be sure we don't introduce mismatches.

* Replace normalize_email utility function with hybrid property
Implements PR feedback, and improves model checks so we also check placeholder badges for valid emails.

* Stop login normalization, and allow multiple results for email search

* Use icontains instead of filter
Per PR suggestion!

* Fix pep8 and unit tests

* Fix email unit tests
For some reason, these weren't caught by Travis.