Ajaxify marking at-door attendees as paid (#2534)
* Ajaxify marking at-door attendees as paid
Halves the number of refreshes on the at-door page by making 'mark as paid' an ajax function. Also brings the manual Stripe form to the at-door reg page as an ajaxified function. Finally, removes some of the scaffolding that is now unnecessary. Fixes https://github.com/magfest/ubersystem/issues/2533.

* Stop using KIOSK_CC_ENABLED for admin-facing controls
Kiosks are attendee-facing. We always want admins to be able to use Stripe, if need be.