Add pending email report (#2623)
* add new pending emails report

this is a daily report sent to STOPs that highlights which
email categories require approval to send. We often fail to
remember to check on this often enough resulting in emails
that go out a few days or weeks after the deadline.  This will
help avoid that issue going forward.

If the volume of emails gets too crazy, it can be turned off
via a config option.  I also structured this so the meat of the
report functionality is in its own function, so it can be exposed
via the API for other uses like Slack notifications

* Updates name of mail pending notification DeamonTask

* Changes pending_emails default value to empty dict (which is falsey)