Add address fields for dealers (#2658)
* Add address fields for dealers
Adapts the 'full address' fields to be used for dealer groups as well as (if configured) attendees. Fixes https://github.com/magfest/ubersystem/issues/2649.

* Tidy up templating logic and fix pep8
I was able to rejigger the way things worked so that fewer templates need to care about the address form.

* Add validation JS back in

* Fix validation whitespace

* Include JS on group forms in all cases
The way the templating logic was set up, the group form might not include necessary JS on the prereg page, depending instead on the attendee regform to include it. However, attendees are only asked for their addresses under certain configurations, so this was a bad idea. This switches the logic around.

* Add zip code to dealer address, add 'copy' checkbox
I forgot to include a zipcode in the actual dealer form, even though I'd made a column. Also adds a 'copy address' checkbox for events that use both an attendee and group address.

* Actually copy address
I straight-up forgot to include this change in the last commit. Whoops!

* Use IDs to make JS is macro less collision-prone

* Replace group.address with group.address1

* Edit migration to deal better with sqlite

I... think this is correct?

* Alter down revision to accommodate new migration

We've got a new migration merged into master from another PR -- this file now references that so things stay linear.

* Fix revision version comment

* Add non-sqlite downgrade