Add duplicate confirmation for admins (#2687)
* Add duplicate confirmation for admins
Fixes https://github.com/magfest/ubersystem/issues/943. We cheat a little in that the badge gets created before we ask for confirmation, but at least this way admins get a chance to immediately undo their mistake.

* Change wording to be more distinct between options

The "edit their info" and "take me where I was going" buttons are similar in terms of code logic, but not really in terms of the UX. We're separating them out (no more button group) and altering the wording to be more clear.