Show prices on pages without a REG_TYPES area (#2689)
* Show prices on pages without a REG_TYPES area
Fixes https://github.com/magfest/ubersystem/issues/2433 by triggering togglePrices(); AFTER the badge_types area has been built, without depending on a click event that doesn't happen on some pages. Also tweaks when prices for the base badge are displayed, effectively turning it off for new group members (because they don't pay any money for their badge, so it's confusing to show them their badge's price) and anyone without a badge cost.

* Also don't show badge prices for registered group members
It's confusing if someone inside a group sees that their base badge price is greater than 0, because the group's price is presented separately. We were hiding it for new group members, but now we'll hide it for confirmed group members too.

* Move togglePrices() out of the BADGE_TYPES.row check

togglePrices should always be called, regardless of whether or not BADGE_TYPES is rendered to the page.