Armor up prereg (#2818)
* Fix "missing parameter" 500 errors for prereg_payment
During the prereg rush this year, we had hundreds of people getting stuck on the POST-only page prereg_payment. When they refreshed the page, they'd get a 500 error for their troubles. This catches that and redirects them to the paid preregistrations page instead so they have a chance of knowing if their badge went through.

* Add defaults for parameters

* heavily armor up the logging around anything to do with attendee payments
- spam as much info as we can into the logs so if there are ever issues, we should be able to better see what's going on
- will be useful for unexpected events like the great registrationing of supermag2018

* Slight tweaks, otherwise everything looks gread, commit() instead of flush()