Pushes the hotel email back by another week
We still don't know when we're sending out the email. But we still don't want it to go out early, so I'm just pushing it back by another week.

For anyone in the general public reading this, don't read too much into it. We still haven't decided, and we will probably be changing this date again several times. Either moving it up or moving it back.