Remove use of cached_property in Charge class
We inherited the use of cached_property for several properties in the Charge class when we hand-updated the class to MAGFest's version. However, there are now some issues with these properties, likely because MAGFest uses a newer version of Sideboard that updates cached_property. This should (hopefully) fix a couple bugs with sending data over to Stripe.