Stop requiring first and last name for placeholder emails (#2905)
* Stop requiring first and last name for placeholder emails
Fixes https://github.com/magfest/ubersystem/issues/2904 by first removing the requirement for placeholder badges to have first and last names, and then modifying all placeholder email templates to only try to print the first name if it in fact exists.

* Replace is_dealer check that was accidentally removed

* Add a. back

* Remove extra `a.staffing` check
When we use the StopsEmail category, we are already checking for the truthiness of `a.staffing`. We don't need to do it again!