Make the intention of the assigning group code clearer
We list several parameters that we weren't actually saving over the 'new' attendee, which made the code look broken. We still want to list the attributes that wouldn't be applied in case their restricted status changes in the future, but phrasing it as deleting attributes from the parameter list makes a lot more sense to humans.