Always give department heads a dept_head_ribbon (#3021)
There's been confusion about the status of the department head ribbon -- we may want to later remove it as a property, but for now we're just automatically adding it to anyone who has the dept_head status. We talked about this a little bit and decided that it's easier to reject a ribbon if you don't want it when you pick up your badge than to ask for a ribbon that's not recorded in the system, so even in the rare case that a department head doesn't want the ribbon, it's okay for the system to force them to have one.