Clean up WatchList page for attendees (#3088)
* Clean up WatchList page for attendees
Fixes https://github.com/magfest/ubersystem/issues/3066. This was meant to be a backend change, but I ended up redoing most of the interface, which https://github.com/magfest/ubersystem/pull/3071 also does. Whoops.

* Move 'active' and 'confirm' actions for watchlist

Now they'll only execute if we have watchlist_id in params, preventing unexpected problems if somehow the watchlist ID is not passed to the page handler.

* pep8