Stop always requiring second_shirt for staff
If one's shirt size was marked, shirt_info_marked was only truthy if either the person in question wasn't staff or if their second_shirt was set -- this meant that for events without 2 shirts, staffers would never be able to move past the 'shirt' step. This fixes that by checking the event's config, much like the shirt template does.