Port 2017 upgrades from MFF (#3190)
* Only show "delete event" button on events that aren't new
We were checking the event ID, but newly instanced events still have IDs, so the button would still show. Clicking the button on a new event would give a 500 error for no row found.

* Allow all_panelists to be overridden in event plugins
MFF wants anyone with a staff ribbon to also be in the list of available panelists, so the definition of panelists has been moved to the SessionMixIn class where it can be overridden.

* Add panel epoch and eschaton
Some events want a different start and end time for their schedule than for their normal epoch/eschaton. This adds that ability.

* Fix flake8 failure