Update API Version and fix config.info call (#3224)
Update API Version and fix config.info call

This change updated the API Version to 1.0, since we made a breaking change to the API Authentication and code written for the 0.1 version is universally broken.

This change also updated the variable YEAR to EVENT_YEAR, tracking an earlier change to the config object.  It both returns the new variable as EVENT_YEAR and returns the legacy variable as YEAR for compatibility with APIs that rewrite their auth code.