Track individual attendees' and groups' share of each StripeTransaction (#3310)
* Track individual attendees' and groups' share of each StripeTransaction
Major overhaul of the stripe transaction and refund system to fix https://jira.magfest.net/browse/MAGDEV-286 and https://jira.magfest.net/browse/MAGDEV-287. We add many-to-many relationship tables that also track each attendee (or group)'s "share" of a transaction.

* Add data migration for multi-model StripeTransactions
This migration hinges on one very important fact: multi-model Stripe transactions ONLY occur when a model is first being created. That means, rather than faffing around and making guesses, we just need to pull the model's `amount_paid` from the Tracking object for its creation. The only major failure point left, then, is that we have to find Attendees and Groups by their name (pulled from the Stripe Transaction's description, the only record of who is in the transaction). This migration gets around that by checking the number of groups and attendees matching a name against the number of times the name occurs in the transaction's description -- if there's a mismatch, it throws an error rather than attempting to create tables for those attendees and groups.

Currently, the migration is in 'test' mode -- it won't commit changes if it detects an error, and downgrading simply destroys the new tables (which the upgrade created, so we're not losing data). This allowed easy testing with my local data, but we should still back up the server before trying this just in case.

* Fix unit tests and handle edge cases found in manual migration tests
I never did recreate the "Key is not present in table" error, but I found many others. Did you know we sometimes create a charge for a model with no `amount_unpaid` on it? I didn't!

* Minor code cleanup

* Allow for Stripe Transactions without attendees
This was actually caused by an import error, but it's a good error to catch.

* Account for common cases found in prod data
People who used ', ' in their names were breaking the migration, along with group leaders who kicked in while buying their group (turns out we don't record that in the Stripe transaction description!).

* Fix that last error
SHIP
IT