Add extendable Guidebook export using Guidebook's "custom list" template (#3389)
* Add extendable Guidebook export using Guidebook's "custom list" template
Fixes https://jira.magfest.net/browse/MAGDEV-309. Many of our exports have to be hardcoded just because our models don't necessarily mirror how we want to display information in Guidebook, so this is about the most flexible/easily extendable way we can do this now. In the future, if we want to add a new model, we simply have to add it to the list in config.py, define the appropriate properties on that model, and in some cases update the export page handler to select the right models for us.

* Improve and fix guidebook exports
We now account for a few more edge cases and have a slightly cleaner page handler. We also have better filenames, with the YYYYMMDD included.

* Use XLS format and handle newlines in data
Fixes https://jira.magfest.net/browse/MAGDEV-373 (hopefully). It also turns out that if a description had a line break in it, our export would just remove everything after the linebreak -- now we handle it properly by replacing \n with <br/>.

* Add option to export zip file of images for Guidebook
Fixes https://jira.magfest.net/browse/MAGDEV-376.

* Switch to using Event model instead of PanelApplication model
Fixes https://jira.magfest.net/browse/MAGDEV-389. We realized that not all panel-events actually have applications, so this changes the logic to just select all events in a designated Panel Room instead.