Let prereg attendees use PromoCodeGroup codes instead of PromoCode codes Alters the lookup_promo_code function to first look for promo codes, then promo code groups -- if it finds the latter and not the former, it takes the first promo code it finds inside the promo code group and applies that to the attendee.
A surprisingly difficult change was to print the group's name on the index page while an attendee was claiming the badge -- attendee.promo_code.group always evaluates to None on that page. I finally found that promo codes were saving with their PromoCodeGroup relationship correctly, but the way we build the Attendee from the session in our Charge class just doesn't include the promo code's group on that specific page. I ended up implementing a somewhat clunky workaround.