Add promo code groups (#3439)
* WIP -- replace group pre-reg with promo code groups
This doesn't work yet -- the name and badges properties we're attaching to the Attendee somehow disappear from the unsaved_preregs attributes once we access them.

* Use Promo Code Groups for pre-reg instead of Groups
When an attendee registers as a group leader, the system now creates a set of promo codes under a Promo Code Group, rather than a Group object. This addresses 1 and 2 of https://jira.magfest.net/browse/MAGDEV-421.

* Let prereg attendees use PromoCodeGroup codes instead of PromoCode codes
Alters the lookup_promo_code function to first look for promo codes, then promo code groups -- if it finds the latter and not the former, it takes the first promo code it finds inside the promo code group and applies that to the attendee.

A surprisingly difficult change was to print the group's name on the index page while an attendee was claiming the badge -- attendee.promo_code.group always evaluates to None on that page. I finally found that promo codes were saving with their PromoCodeGroup relationship correctly, but the way we build the Attendee from the session in our Charge class just doesn't include the promo code's group on that specific page. I ended up implementing a somewhat clunky workaround.

* Add PromoCode group management page
Also adds cost tracking to promo codes so we know what price they were bought at.

* Fix migration order
For some reason the latest migration was pointing to the first revision instead of the prior one. Whoops!

* Create promo code group management page for group leaders
This is the last big chunk of work for adding the promo code group feature (https://jira.magfest.net/browse/MAGDEV-421).

* Clean up loose ends except unit tests
Adds an email I forgot, reverts some code that got changed unnecessarily, and fixes flake8 errors. Unit test fixes are pending as my dev environment will not currently run them.