Stop re-installing some libraries
We're not sure what the rhyme or reason for this is, but our deploy process is reinstalling older versions of some libraries on top of newer versions, and tripping up a shared filesystem bug (https://github.com/magfest/reggie-formula/issues/16). This changes some of the libraries we found that did this, though this likely doesn't catch them all.