Retool how Staff and Event shirts interact to allow more flexibility The way we handled staff and event shirts was not flexible enough for MAGWest's purposes. I've replaced our `second_shirt` column with `num_event_shirts`, which allows events to add configuration for any number of shirts along with c.SHIRTS_PER_STAFFER. For example, an event can now have three shirts per staffer and let them replace either one or two (or none) of them, using just config.
We do still assume that staffers (and not volunteers) are the only ones who get this selection, but this can be overridden using two properties (which you'll see shortly with MAGWest). The template for selecting shirt type/size is also still using Super MAGFest text, which we may want to move later or just override for other events that want to use this event-shirt-replacing selection.