Convert Attendee and Group forms to use modular fields (#3464)
* Template overhaul WIP
Far from finished, but I wanted to put this up for review as a sanity check before I dig myself too deep

* Convert admin form to modular fields
Clean-up and testing still needed. This is the bulk of the work in building modular attendee fields for our forms.

* Convert group fields to modular fields
This should wrap up work on this step of the template overhaul. All attendee and group fields should now be re-implemented as modular fields that can dynamically be set to read-only.

* Incorporate recent updates to attendee and group forms