Stop resetting ANY boolean values on the admin reg form (#3474)
* Stop resetting ANY boolean values on the admin reg form
Fixes https://jira.magfest.net/browse/MAGDEV-534. Also (hopefully) fixes any other 'reset' bug for this form, except for `got_swadge` (which is supposedly not meant to be permanent so I'm not adding new code) and any columns that come from event plugins (we'll need to fix those in their respective plugins). Checkboxes always pass either a string, or no value at all -- we tell our page handler to process them by telling it which fields are boolean, so it can interpret a value as "Yes" and a lack of value as "No." However, we were telling it to check every boolean column that an Attendee object could have, and many of those values were not actually represented on our admin form. Now they are.

* Re-enable accessibility checkbox on non-admin pages