Combine MIVS rounds one and two Fixes https://jira.magfest.net/browse/MAGDEV-544. Also cleans up all the old MIVS emails.
The following email templates should be reviewed by the MIVS team:
video_broken.txt -- The original text stated that a game with a broken video would be declined. I updated it to remove references to round 1, but I'm not sure if a game that can be played but whose video is broken would still be declined.
checklist_new_item.txt -- There was never any text submitted for this email from last year!