Fix performance issues for Hotel Export (#3543)
* Fix performance issues for Hotel Export
We found out at the last minute that the new format for the hotel export caused a 504 error, and had to make an emergency commit (https://github.com/magfest/ubersystem/commit/d64bde9e23ec25b57d4eeae1790331b5c85332c4) to get the export out in time. We still want that format though, so this is a version of the export that has the right format and didn't cause 504 error on my local system (I added 10k placeholder attendees to test).

Also fixes a bug that caused 'late' comped badges to not actually be included in the export. Whoops.

* Improve performance even more
We've replaced the multiple queries with a dictionary lookup. This has also gotten us around whatever problem is happening when trying to query on first name.