Fix API config.info timezone (#3583)
* Fix API config.info timezone

I previously added EVENT_TIMEZONE to the list returned by config.info.  Upon trying to actually use it, the JSON encoder returns an error because this is a TZINFO object and it doesn't know how to handle it.  This change adds code to convert this to a string equivalent before JSON encoding.