Refactor site map and access groups page list We were using two separate, slightly different chunks of complex code to introspect the application and build a list of site sections and pages. We were also not caching the site map, causing it to take multiple seconds to load.
This combines the two methods into one cached property that returns all the data structures we need to build both the Site Map and the list of pages and sections for the Access Groups page. This is safe to make fully cached, since the app's structure does not change dynamically.