Preserve unassigned group badges' staffing status (#3670)
* Preserve unassigned group badges' staffing status
Fixes https://jira.magfest.net/browse/MAGDEV-789. If someone claims an unassigned group badge, even if that badge was 'staffing' the form was defaulting to an unchecked staffing checkbox. People should never be able to opt out of staffing when claiming a Contractor or Staff badge in a group, so we hide the checkbox and force staffing to be true.

* Clean up extraneous change