Allow adding numbers to attendees during at-con mode (#3704)
Fixes https://github.com/magfest/ubersystem/issues/3703. Normally we zero out a badge number if an attendee doesn't need one. This doesn't make sense for at-the-con, when we may need to assign badge numbers ahead of time. So if we're at the con, we don't zero out the number. We also don't auto-assign badge numbers -- we throw an error if you try to check in someone without one. Also also, we no longer make "save and check-in" the default action when you press enter on the form.