Don't mark actions as badge shifts if badge shifting is turned off
The tracking database code is so naive that admins changing badge numbers on purpose is marked as an automatic badge shift, even if the shift custom badge setting is False. This makes it at least smart enough to not scare people reviewing the tracking table into thinking that have shift_custom_badges set wrong.