Provide link to reset default printer IDs Admins may sometimes want to reset their default printer IDs. Before you would have to log out and back in, which is tedious and also resets your reg station #. Now you can use a link to quickly null your default printer IDs so you can set them during the next check-in.
Note that because printer IDs get set via an AJAX call, this link is always displayed even if printer IDs aren't set when the page is load. That way, if someone mistypes a printer ID, they can just close the check-in modal and click the link without having to refresh the page.