Fix tracking links and get around bad tracking Fixes https://jira.magfest.net/browse/MAGDEV-1035, mostly. The tracking table links each entry to other models by way of `links` and automatically builds this list by looking at foreign keys. However, since we started tracking the creator of Attendees and Groups, admins would be linked to every single attendee and group they ever made, and all THOSE attendee/groups' changes would be linked to that admin.
The link is fixed, but there's no clean way to fix the data, so we also filter out tracking results for any attendees that are linked to the admin. This doesn't fix groups' changes appearing, but we don't want to remove tracking data for those since we'll also remove tracking data for groups you're the leader of. That is probably the best we can do for this year.