Update Badge Claim Email
This is my first pull request for these emails, I tried to make it sound more like the Super/West badge claim emails, since this seems to be an older template. Please let me know if I'm doing this wrong (also I might need to re-do this one because we don't know the Covid policy yet).