Fix dealer editable fields and add locking fields Our old way to control when dealers could edit fields was all over the place -- this replaces that with a simple list of which statuses dealers can edit their information in.
To implement this, we've also adding a way to define fields that are locked if you are not an admin. This is much cleaner over the old method of adding validations for each field. If someone tries to edit a field that's locked, we log it and simply don't set the field to the new value.