Fix several issues with upgrading badges I had broken receipt updating for attendees while working on receipt updating for groups, so now the logic is more explicit about which instance we're checking. Also adds safeguards against upgrading attendees if no receipt items are generated.
Also fixes an issue when someone's upgrade was undone but their payment not cancelled by an admin -- the system now automatically cancels incomplete payments if the attendee doesn't owe anything. Also also, validations for upgrades weren't actually running, so we've switched over to the new validation system.

Finally, it turns out that InputRequired doesn't actually do what we want -- WTForms recommends it over DataRequired because I guess they don't imagine you not rendering all fields in a form sometimes -- so this converts InputRequired validations on all fields that get saved to a model to DataRequired so that existing modal data will prevent a validation error if that field doesn't happen to be rendered.