Make assigned departments editable again It turns out the code adds assigned departments through a relationship that had been marked as viewonly. This removes the viewonly marker, which is the quick fix.
The real fix is to adjust how we edit relationships like this -- the heavy abstraction of routing everything through the set_relation_ids functions means that it's difficult to tell which relationships can be made viewonly and which can't. We already have to define setter functions anyway, we might as well require that you specify what class and relationship should be used.