Add kiosk operation for merch booth page Adds a new access control where a logged-in admin can launch a 'kiosk' for a page. This changes the login from an 'admin' login to a 'supervisor' login and then prompts for a badge number to log in the kiosk 'operator.' Both the operator and supervisor's names are attached to all changes logged in the tracking table.
Right now this is only set up for /merch_admin/index but it is definitely possible to extend this to other areas. For security reasons, admin and supervisor logins are treated as mutually exclusive -- launching a kiosk logs you out of your admin account, and logging back into your admin account logs out the kiosk.

Also adds a fix for certain types of decorators causing AJAX functions to fail silently.