Create list of schedule changes for Guidebook Fixes https://magfest.atlassian.net/browse/MAGDEV-1299 by adding a report page that tracks all guidebook-related items' last updated date vs their last sync date. In order to show exactly what updated, we also store a JSONized blob of the item's current state. Also fixes https://magfest.atlassian.net/browse/MAGDEV-1341 by cleaning up what we offer for export and getting rid of a couple old schedule exports meant for Guidebook.
If you want to skip marking individual items you can also mark all items in a custom list (or the schedule) as synced, which is also automatically done when you export data. This method of syncing skips any items that were uploaded since you loaded the schedule_reports index page (or since the export started).

This does NOT automatically sync any items -- for now, admins must manually copy the data over to Guidebook. The scheme in this PR for last_synced should be followed for other integrations when possible (a single key for the datetime, and a key under last_synced['data'] for the json data).

The queries involved in this feature were tested against prod -- accordingly, we're adding the 'track' column to the Events to avoid each event traversing over its applications to build its guidebook data dict.

Finally, due to how images are handled for guests, there's some minor edge-case bugs -- we need to change guest images anyway to be able to accommodate both a header and a thumbnail image, so those can be fixed then. Chiefly, if someone updates their bio pic with the same filename, there is no way for the system to know that they did so.