Let groups add existing badges Gives group leaders the ability to add existing badges by entering their ID, first name, and last name. This is disabled by default via the "can_add_existing_badges" property -- we're using a DB property rather than a config option as it's likely any event that uses it will want to enable it only for specific kinds of groups.
Also fixes a fairly serious bug where a badge that was not paid for by a group could be unset and then reassigned... yikes. Similarly, badges can't be unset if they have an admin account or are a department head.

Also also, breaks up the group members template a bit more -- this was originally to allow override via plugin, but I ended up putting the functionality in the main repo since it relies on page handlers.